moved LanguageCodeFormat to types.ts
