+++ source/languagedetect.ts
@@ -1,7 +1,6 @@
 import LanguageDetector from 'languagedetect';
-import {Language} from './types.js';
+import {Language, LanguageCodeFormat} from './types.js';
 import {languageCodeToName} from './utils.js';
-import {LanguageCodeFormat} from './index.js';
 
 export const detect = (
 	text: string,
