+++ source/types.ts
@@ -2,3 +2,8 @@ export type Language = {
 	languageCode: string;
 	languageName: string;
 };
+
+export enum LanguageCodeFormat {
+	TwoLetter = 'twoLetter',
+	ThreeLetter = 'threeLetter',
+}
